add crud for post

+++ post-service/src/main/java/com/root14/postservice/service/PostService.java
@@ -7,6 +7,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.awt.*;
+import java.util.List;
+import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
 public class PostService {
@@ -14,10 +18,7 @@ public class PostService {
 
     public ResponseEntity<?> addPost(AddPostDto addPostDto, String authenticatedUserId) {
         try {
-            Post post = Post.builder()
-                    .content(addPostDto.getContent())
-                    .authorId(authenticatedUserId)
-                    .build();
+            Post post = Post.builder().content(addPostDto.getContent()).authorId(authenticatedUserId).build();
             postRepository.save(post);
             return ResponseEntity.ok().build();
         } catch (Exception e) {
@@ -25,12 +26,73 @@ public class PostService {
         }
     }
 
-    public void deletePost(String id) {
-        postRepository.deleteById(id);
+    public ResponseEntity<?> deleteByPostId(String id, boolean softDelete) {
+        try {
+            if (softDelete) {
+                Optional<Post> post = postRepository.findById(id);
+                if (post.isPresent()) {
+                    post.get().setEnabled(false);
+                    postRepository.save(post.get());
+                    return ResponseEntity.ok().build();
+                } else {
+                    return ResponseEntity.notFound().build();
+                }
+            } else {
+                try {
+                    postRepository.deleteById(id);
+                    return ResponseEntity.ok().build();
+                } catch (Exception e) {
+                    return ResponseEntity.notFound().build();
+                }
+            }
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body("Something went wrong." + e.getMessage());
+        }
     }
 
-    public Post getPost(String id) {
-        return postRepository.findById(id).orElseThrow();
+    //todo do not return object raw
+    public ResponseEntity<?> getPostByPostId(String id) {
+        try {
+            Optional<Post> post = postRepository.findById(id);
+
+            if (post.isPresent()) {
+                return ResponseEntity.ok().body(post.get());
+            } else {
+                return ResponseEntity.notFound().build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body("Something went wrong." + e.getMessage());
+        }
+    }
+
+    //todo do not return object raw
+    public ResponseEntity<?> getPostByAuthorId(String authorId) {
+        try {
+            Optional<List<Post>> optionalPostList = postRepository.findPostByAuthorId(authorId);
+
+            if (optionalPostList.isPresent()) {
+                return ResponseEntity.ok().body(optionalPostList.get());
+            } else {
+                return ResponseEntity.notFound().build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body("Something went wrong." + e.getMessage());
+        }
+    }
+
+    public ResponseEntity<?> updatePostByPostId(String postId, String content) {
+        try {
+            Optional<Post> post = postRepository.findById(postId);
+            if (post.isPresent()) {
+                post.get().setContent(content);
+                postRepository.save(post.get());
+                return ResponseEntity.ok().build();
+            } else {
+                return ResponseEntity.notFound().build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body("Something went wrong." + e.getMessage());
+        }
     }
